Rewrite `take` to use one less `unsafe` trick
The main logic of `take` is rewritten in the same manner as `insert` was
rewritten in the previous commit: use the call stack instead of a `Vec`
to hold the mutable references to all the parent pointers.

The remaining uses of `unsafe` are in the case where the removed node
has both a left and a right child. That code is unchanged for now,
except for some renamed variables.